+++ {{cookiecutter.project_name}}/tasks.py
@@ -109,9 +109,9 @@ def flake8(c):
 
 
 @task()
-def safety(c):
+def security(c):
     # type: (Context) -> None
-    """Run safety."""
+    """Run security related checks."""
     _run(
         c,
         "poetry export --with dev --format=requirements.txt --without-hashes | "
@@ -119,7 +119,7 @@ def safety(c):
     )
 
 
-@task(pre=[flake8, safety, call(format_, check=True)])
+@task(pre=[flake8, security, call(format_, check=True)])
 def lint(c):
     # type: (Context) -> None
     """Run all linting."""