+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Renamed invoke safety task to invoke security.
+
 ## [2.2.1] - 2023-05-25
 
 ### Changed